pyproject.toml
[project]
name = "reviewer-reflex"
version = "0.1.0"
description = "AI-powered GitHub PR reviewer built with Reflex"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "httpx>=0.28.1",
    "openai>=1.60.0",
    "python-dotenv>=1.2.1",
    "reflex>=0.8.24.post1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.10",
]

[tool.ruff.lint]
extend-select = ["I"]
